Move GetStartOnSystemStartup and SetStartOnSystemStartup to GUI code
