+++ src/qt/bitcoin.cpp
@@ -232,6 +232,10 @@ int main(int argc, char *argv[])
 
     try
     {
+        // Regenerate startup link, to fix links to old versions
+        if (GUIUtil::GetStartOnSystemStartup())
+            GUIUtil::SetStartOnSystemStartup(true);
+
         BitcoinGUI window;
         guiref = &window;
         if(AppInit2(argc, argv))
